requirements.txt
streamlit==1.32.2

pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2

matplotlib==3.8.2
pyyaml==6.0.1

tqdm>=4.64,<5.0
joblib>=1.2,<2.0
